--- apps/store-integration/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "integration-tests",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "test": "vitest run"
-  },
-  "dependencies": {
-    "@aws-sdk/client-dynamodb": "^3.258.0",
-    "@equinox-js/core": "workspace:*",
-    "@equinox-js/dynamo-store": "workspace:*",
-    "@equinox-js/message-db": "workspace:*",
-    "@opentelemetry/api": "^1.4.0",
-    "@opentelemetry/sdk-trace-base": "^1.9.0",
-    "@opentelemetry/sdk-trace-node": "^1.9.0",
-    "pg": "^8.8.0",
-    "ramda": "^0.28.0",
-    "vitest": "^0.34.1"
-  },
-  "devDependencies": {
-    "@types/node": "^18.11.18",
-    "@types/pg": "^8.6.6",
-    "@types/ramda": "^0.28.21"
-  }
-}
