+++ packages/dynamo-store/dynamo-store/package.json
@@ -14,16 +14,19 @@
     "build": "tsup src/index.ts --format esm,cjs --dts --clean --sourcemap"
   },
   "devDependencies": {
+    "@opentelemetry/sdk-trace-base": "^1.9.0",
+    "@opentelemetry/sdk-trace-node": "^1.9.0",
     "@types/node": "^18.11.18",
     "tsup": "^7.2.0",
     "typescript": "^5.2.2",
-    "vitest": "^0.25.3"
+    "vitest": "^0.34.3"
   },
   "dependencies": {
     "@aws-sdk/client-dynamodb": "^3.258.0",
     "@aws-sdk/smithy-client": "^3.257.0",
     "@aws-sdk/types": "^3.257.0",
     "@equinox-js/core": "workspace:*",
+    "@equinox-js/test-domain": "workspace:*",
     "@opentelemetry/api": "^1.4.0"
   }
 }
