+++ packages/message-db/message-db/package.json
@@ -18,6 +18,9 @@
     "./dist"
   ],
   "devDependencies": {
+    "@equinox-js/test-domain": "workspace:*",
+    "@opentelemetry/sdk-trace-base": "^1.9.0",
+    "@opentelemetry/sdk-trace-node": "^1.9.0",
     "@types/node": "^18.11.18",
     "@types/pg": "^8.6.6",
     "tsup": "^7.2.0",
