+++ packages/message-db/message-db/tsconfig.json
@@ -17,5 +17,5 @@
     "skipLibCheck": true,
     "strict": true
   },
-  "exclude": ["node_modules"]
+  "exclude": ["node_modules", "dist"]
 }
