+++ packages/test-domain/src/index.ts
@@ -1,0 +1,2 @@
+export * as Cart from "./domain/Cart.js"
+export * as ContactPreferences from "./domain/ContactPreferences.js"
